[refactor] UsageViewModel: split function into GenerateUsageInsightsUseCase
