+++ composeApp/src/commonMain/kotlin/com/rwmobi/kunigami/di/ViewModelModule.kt
@@ -7,6 +7,7 @@
 
 package com.rwmobi.kunigami.di
 
+import com.rwmobi.kunigami.domain.usecase.GenerateUsageInsightsUseCase
 import com.rwmobi.kunigami.ui.viewmodels.AccountViewModel
 import com.rwmobi.kunigami.ui.viewmodels.AgileViewModel
 import com.rwmobi.kunigami.ui.viewmodels.TariffsViewModel
@@ -20,6 +21,7 @@ val viewModelModule = module {
             syncUserProfileUseCase = get(),
             getConsumptionAndCostUseCase = get(),
             getTariffSummaryUseCase = get(),
+            generateUsageInsightsUseCase = GenerateUsageInsightsUseCase(),
             dispatcher = get(named("DefaultDispatcher")),
         )
     }
